# swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"FieldErrors": {
				"properties": {},
				"type": "object",
				"additionalProperties": {
					"properties": {
						"value": {},
						"message": {
							"type": "string"
						}
					},
					"required": [
						"message"
					],
					"type": "object"
				}
			},
			"ErrorDetails": {
				"anyOf": [
					{
						"properties": {
							"folderName": {
								"type": "string"
							},
							"userId": {
								"type": "string"
							},
							"folderId": {
								"type": "string"
							}
						},
						"type": "object"
					},
					{
						"properties": {
							"imageUrl": {
								"type": "string"
							},
							"imageId": {
								"type": "string"
							}
						},
						"type": "object"
					},
					{
						"properties": {
							"userId": {
								"type": "string"
							},
							"challengeId": {
								"type": "string"
							}
						},
						"type": "object"
					},
					{
						"properties": {
							"longitude": {
								"type": "number",
								"format": "double"
							},
							"latitude": {
								"type": "number",
								"format": "double"
							}
						},
						"type": "object"
					},
					{
						"properties": {
							"reason": {
								"type": "string"
							}
						},
						"type": "object"
					},
					{
						"properties": {
							"searchKeyword": {
								"type": "string"
							}
						},
						"type": "object"
					},
					{
						"$ref": "#/components/schemas/FieldErrors"
					},
					{
						"properties": {
							"extension": {
								"type": "string"
							}
						},
						"type": "object"
					},
					{
						"properties": {
							"imageId": {
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							"folderId": {
								"type": "string"
							}
						},
						"type": "object"
					}
				],
				"nullable": true
			},
			"ITsoaErrorResponse": {
				"properties": {
					"resultType": {
						"type": "string"
					},
					"error": {
						"properties": {
							"data": {
								"$ref": "#/components/schemas/ErrorDetails"
							},
							"reason": {
								"type": "string"
							},
							"errorCode": {
								"type": "string"
							}
						},
						"required": [
							"data",
							"reason",
							"errorCode"
						],
						"type": "object"
					},
					"success": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseTagListFromImage": {
				"properties": {
					"tags": {
						"items": {
							"properties": {
								"tagCategory": {
									"properties": {
										"tagType": {
											"type": "string"
										},
										"id": {
											"type": "string"
										}
									},
									"required": [
										"tagType",
										"id"
									],
									"type": "object"
								},
								"content": {
									"type": "string"
								}
							},
							"required": [
								"tagCategory",
								"content"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"tags"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ITsoaSuccessResponse_ResponseTagListFromImage_": {
				"properties": {
					"resultType": {
						"type": "string"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"$ref": "#/components/schemas/ResponseTagListFromImage"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseGetImageListFromTag": {
				"properties": {
					"images": {
						"items": {
							"properties": {
								"mediaId": {
									"type": "string"
								},
								"id": {
									"type": "string"
								}
							},
							"required": [
								"mediaId",
								"id"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"images"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ITsoaSuccessResponse_ResponseGetImageListFromTag_": {
				"properties": {
					"resultType": {
						"type": "string"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"$ref": "#/components/schemas/ResponseGetImageListFromTag"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseCreateImage": {
				"properties": {
					"imageId": {
						"type": "string"
					}
				},
				"required": [
					"imageId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ITsoaSuccessResponse_ResponseCreateImage_": {
				"properties": {
					"resultType": {
						"type": "string"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"$ref": "#/components/schemas/ResponseCreateImage"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"MemoTextImageListResponseDto": {
				"properties": {
					"folderId": {
						"type": "string"
					},
					"folderName": {
						"type": "string"
					},
					"imageText": {
						"type": "string"
					},
					"images": {
						"items": {
							"properties": {
								"imageUrl": {
									"type": "string"
								},
								"imageId": {
									"type": "string"
								}
							},
							"required": [
								"imageUrl",
								"imageId"
							],
							"type": "object"
						},
						"type": "array",
						"nullable": true
					}
				},
				"required": [
					"folderId",
					"folderName",
					"imageText",
					"images"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ITsoaSuccessResponse_MemoTextImageListResponseDto_": {
				"properties": {
					"resultType": {
						"type": "string"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"$ref": "#/components/schemas/MemoTextImageListResponseDto"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BodyToMemoImagesToMove": {
				"properties": {
					"targetFolderId": {
						"type": "string"
					},
					"imageId": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"targetFolderId",
					"imageId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"sessionAuth": {
				"type": "apiKey",
				"in": "cookie",
				"name": "connect.sid"
			}
		}
	},
	"info": {
		"title": "sweepic-server",
		"version": "1.0.0",
		"description": "Sweepic 프로젝트",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"servers": [
		{
			"url": "http://localhost:3000",
			"description": "Sweepic local server"
		},
		{
			"url": "http://3.37.137.212:3000",
			"description": "Sweepic server"
		}
	],
	"paths": {
...
		"/tags/images/{mediaId}": {
			"get": {
				"operationId": "GetTagListFromImage",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaSuccessResponse_ResponseTagListFromImage_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "VAL-001",
												"reason": "Validation Error",
												"data": {
													"mediaId": {
														"message": "invalid float",
														"value": "a"
													}
												}
											},
											"success": null
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "TAG-001",
												"reason": "태그가 없습니다.",
												"data": null
											},
											"success": null
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "SER-001",
												"reason": "내부 서버 오류입니다.",
												"data": {
													"reason": "에러원인 메시지"
												}
											},
											"success": null
										}
									},
									"Example 2": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "DB-001",
												"reason": "DB 에러 입니다.",
												"data": {
													"reason": "에러원인 메시지"
												}
											},
											"success": null
										}
									}
								}
							}
						}
					}
				},
				"description": "사용자가 mediaId를 제공하면 해당 이미지와 연관된 태그들의 정보를 제공합니다.",
				"summary": "이미지의 태그 조회",
				"tags": [
					"Tag"
				],
				"security": [],
				"parameters": [
					{
						"description": "이미지의 고유 mediaId",
						"in": "path",
						"name": "mediaId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/images": {
			"get": {
				"operationId": "GetImageListFromTag",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaSuccessResponse_ResponseGetImageListFromTag_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "VAL-001",
												"reason": "Validation Error",
												"data": {
													"tag": {
														"message": "\"tag\" is required"
													}
												}
											},
											"success": null
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "PHO-404",
												"reason": "<인하대학교> 태그에 해당하는 사진이 존재하지 않습니다.",
												"data": null
											},
											"success": null
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "SER-001",
												"reason": "내부 서버 오류입니다.",
												"data": {
													"reason": "에러원인 메시지"
												}
											},
											"success": null
										}
									},
									"Example 2": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "DB-001",
												"reason": "DB 에러 입니다.",
												"data": {
													"reason": "에러원인 메시지"
												}
											},
											"success": null
										}
									}
								}
							}
						}
					}
				},
				"description": "사용자가 tag를 입력하면 해당 태그에 속하는 이미지들을 가져옵니다.",
				"summary": "태그기반 이미지 검색",
				"tags": [
					"Image"
				],
				"security": [],
				"parameters": [
					{
						"description": "태그 내용",
						"in": "query",
						"name": "tag",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateImage",
				"responses": {
					"201": {
						"description": "CREATED",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaSuccessResponse_ResponseCreateImage_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "VAL-001",
												"reason": "Validation Error",
												"data": {
													"mediaId": {
														"message": "invalid float",
														"value": "a"
													}
												}
											},
											"success": null
										}
									},
									"Example 2": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "PHO-400",
												"reason": "사진 데이터가 유효하지 않습니다.",
												"data": {
													"reason": "이미 휴지통에 있는 사진입니다"
												}
											},
											"success": null
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "SER-001",
												"reason": "내부 서버 오류입니다.",
												"data": {
													"reason": "에러원인 메시지"
												}
											},
											"success": null
										}
									},
									"Example 2": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "DB-001",
												"reason": "DB 에러 입니다.",
												"data": {
													"reason": "에러원인 메시지"
												}
											},
											"success": null
										}
									}
								}
							}
						}
					}
				},
				"description": "mediaId에 해당하는 사진데이터가 존재하면 해당 사진데이터의 imageId를 보여줍니다.(조회)\r\n\r\nmediaId에 해당하는 사진데이터가 존재하지 않는다면 사진데이터를 생성하고 생성한 사진데이터의 imageId를 보여줍니다.(생성)",
				"summary": "이미지 데이터 생성 및 조회",
				"tags": [
					"Image"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "미디어ID, 사진 생성일",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"timestamp": {
										"type": "string",
										"format": "date-time"
									},
									"mediaId": {
										"type": "string"
									}
								},
								"required": [
									"timestamp",
									"mediaId"
								],
								"type": "object",
								"description": "미디어ID, 사진 생성일"
							}
						}
					}
				}
			}
		},
		"/memo/folders/{folderId}/images": {
			"patch": {
				"operationId": "HandlerMemoImageMove",
				"responses": {
					"200": {
						"description": "사진 이동 성공 응답",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaSuccessResponse_MemoTextImageListResponseDto_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "SUCCESS",
											"error": null,
											"success": {
												"folderId": "1",
												"folderName": "string",
												"imageText": "string",
												"images": [
													{
														"imageId": "1",
														"imageUrl": "string"
													}
												]
											}
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "유효하지 않은 데이터 조회 에러",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "MEM-400",
												"reason": "메모 사진 이동 중 오류가 발생했습니다.",
												"data": {
													"folderId": "1",
													"imageId": [
														"1"
													]
												}
											},
											"success": null
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않은 데이터 조회 에러",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "FOL-404",
												"reason": "해당 폴더를 찾을 수 없습니다.",
												"data": {
													"folderId": "1"
												}
											},
											"success": null
										}
									},
									"Example 2": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "PHO-404",
												"reason": "해당 사진 데이터가 없습니다.",
												"data": {
													"imageId": "1"
												}
											},
											"success": null
										}
									}
								}
							}
						}
					}
				},
				"description": "특정 폴더의 사진을 이동하는 API입니다.",
				"summary": "사진 이동 API",
				"tags": [
					"memo-image-controller"
				],
				"security": [],
				"parameters": [
					{
						"description": "현재 폴더 ID",
						"in": "path",
						"name": "folderId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "이동할 폴더 ID, 이동할 사진 배열",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BodyToMemoImagesToMove",
								"description": "이동할 폴더 ID, 이동할 사진 배열"
							}
						}
					}
				}
			}
		},
...
	}
}